Fix #327 on ripples.js and improve structure
I've fixed #327 and I also passed the plugin to a jQuery Boilerplate structure. I only made changes on the structure and not in the functions themselves, so the  code could be a little bit neater, cleaner and intuitive.

Now the plugin can be accessed alone, by doing: 
```javascript
$(selector).ripples();
```

Note: I've removed the default elements targeted by the plugin, for IMHO I think that should be in the initialization, and not in the plugin itself.